@@ -23,10 +23,14 @@ import net.tharow.tantalum.launchercore.auth.IAuthListener;
 import net.tharow.tantalum.launchercore.auth.IUserType;
 import net.tharow.tantalum.launchercore.modpacks.*;
 import net.tharow.tantalum.launchercore.modpacks.packinfo.CombinedPackInfo;
+import net.tharow.tantalum.launchercore.modpacks.sources.IAuthoritativePackSource;
 import net.tharow.tantalum.launchercore.modpacks.sources.IPackSource;
 import net.tharow.tantalum.launchercore.modpacks.sources.NameFilterPackSource;
 import net.tharow.tantalum.platform.IPlatformApi;
 import net.tharow.tantalum.platform.IPlatformSearchApi;
+import net.tharow.tantalum.platform.PlatformPackInfoRepository;
+import net.tharow.tantalum.platform.http.HttpPlatformApi;
+import net.tharow.tantalum.platform.http.HttpPlatformSearchApi;
 import net.tharow.tantalum.platform.io.PlatformPackInfo;
 import net.tharow.tantalum.platform.packsources.SearchResultPackSource;
 import net.tharow.tantalum.platform.packsources.SinglePlatformSource;
@@ -34,6 +38,7 @@ import net.tharow.tantalum.rest.RestfulAPIException;
 import net.tharow.tantalum.rest.io.PackInfo;
 import net.tharow.tantalum.solder.ISolderApi;
 import net.tharow.tantalum.solder.ISolderPackApi;
+import net.tharow.tantalum.solder.http.HttpSolderApi;
 import net.tharow.tantalum.ui.controls.TintablePanel;
 import net.tharow.tantalum.ui.controls.WatermarkTextField;
 import net.tharow.tantalum.ui.controls.borders.RoundBorder;
@@ -44,6 +49,8 @@ import net.tharow.tantalum.ui.controls.list.SimpleScrollbarUI;
 import net.tharow.tantalum.launcher.ui.controls.modpacks.ModpackWidget;
 import net.tharow.tantalum.launchercore.image.ImageRepository;
 import net.tharow.tantalum.utilslib.DesktopUtils;
+import net.tharow.tantalum.utilslib.Utils;
+import org.jboss.logging.Logger;
 
 import javax.swing.*;
 import javax.swing.Timer;
@@ -62,6 +69,7 @@ import java.net.URISyntaxException;
 import java.net.URLDecoder;
 import java.net.URLEncoder;
 import java.util.*;
+import java.util.logging.Level;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -78,6 +86,7 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
     private JScrollPane scrollPane;
     private ModpackInfoPanel modpackInfoPanel;
     private LauncherFrame launcherFrame;
+    private JComboBox<String> source;
     private JTextField filterContents;
     private FindMoreWidget findMoreWidget;
 
@@ -109,12 +118,7 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
         siteRegex = Pattern.compile("^([a-zA-Z0-9-]+)\\.\\d+$");
 
         findMoreWidget = new FindMoreWidget(resources);
-        findMoreWidget.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                DesktopUtils.browseUrl(findMoreUrl);
-            }
-        });
+        findMoreWidget.addActionListener(e -> DesktopUtils.browseUrl(findMoreUrl));
 
         relocalize(resources);
     }
@@ -238,12 +242,9 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
         if (modpack.hasRecommendedUpdate()) {
             widget.setToolTipText(resources.getString("launcher.packselector.updatetip"));
         }
-        widget.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                if (e.getSource() instanceof ModpackWidget)
-                    selectWidget((ModpackWidget)e.getSource());
-            }
+        widget.addActionListener(e -> {
+            if (e.getSource() instanceof ModpackWidget)
+                selectWidget((ModpackWidget)e.getSource());
         });
 
         if (widget.getModpack().isSelected()) {
@@ -258,14 +259,11 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
         rebuildUI();
 
         if (selectedWidget != null) {
-            EventQueue.invokeLater(new Runnable() {
-                @Override
-                public void run() {
-                    if (widget == selectedWidget)
-                        selectWidget(widget);
-                    else
-                        selectedWidget.scrollRectToVisible(new Rectangle(selectedWidget.getSize()));
-                }
+            EventQueue.invokeLater(() -> {
+                if (widget == selectedWidget)
+                    selectWidget(widget);
+                else
+                    selectedWidget.scrollRectToVisible(new Rectangle(selectedWidget.getSize()));
             });
         }
     }
@@ -289,21 +287,18 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
         if (selectedWidget == null || selectedWidget.getModpack() == null || !allModpacks.containsKey(selectedWidget.getModpack().getName())) {
             java.util.List<ModpackWidget> sortedPacks = new LinkedList<ModpackWidget>();
             sortedPacks.addAll(allModpacks.values());
-            Collections.sort(sortedPacks, new Comparator<ModpackWidget>() {
-                @Override
-                public int compare(ModpackWidget o1, ModpackWidget o2) {
-                    int priorityCompare = (new Integer(o2.getModpack().getPriority())).compareTo(new Integer(o1.getModpack().getPriority()));
-                    if (priorityCompare != 0)
-                        return priorityCompare;
-                    else if (o1.getModpack().getDisplayName() == null && o2.getModpack().getDisplayName() == null)
-                        return 0;
-                    else if (o1.getModpack().getDisplayName() == null)
-                        return -1;
-                    else if (o2.getModpack().getDisplayName() == null)
-                        return 1;
-                    else
-                        return o1.getModpack().getDisplayName().compareToIgnoreCase(o2.getModpack().getDisplayName());
-                }
+            Collections.sort(sortedPacks, (o1, o2) -> {
+                int priorityCompare = (new Integer(o2.getModpack().getPriority())).compareTo(new Integer(o1.getModpack().getPriority()));
+                if (priorityCompare != 0)
+                    return priorityCompare;
+                else if (o1.getModpack().getDisplayName() == null && o2.getModpack().getDisplayName() == null)
+                    return 0;
+                else if (o1.getModpack().getDisplayName() == null)
+                    return -1;
+                else if (o2.getModpack().getDisplayName() == null)
+                    return 1;
+                else
+                    return o1.getModpack().getDisplayName().compareToIgnoreCase(o2.getModpack().getDisplayName());
             });
             if (sortedPacks.size() > 0) {
                 selectWidget(sortedPacks.get(0));
@@ -341,29 +336,23 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
                         if (infoToUse != null)
                             refreshWidget.getModpack().setPackInfo(infoToUse);
 
-                        EventQueue.invokeLater(new Runnable() {
-                            @Override
-                            public void run() {
-                                if (modpackInfoPanel != null)
-                                    modpackInfoPanel.setModpackIfSame(refreshWidget.getModpack());
+                        EventQueue.invokeLater(() -> {
+                            if (modpackInfoPanel != null)
+                                modpackInfoPanel.setModpackIfSame(refreshWidget.getModpack());
 
-                                if (refreshWidget.getModpack().hasRecommendedUpdate()) {
-                                    refreshWidget.setToolTipText(resources.getString("launcher.packselector.updatetip"));
-                                } else {
-                                    refreshWidget.setToolTipText(null);
-                                }
-
-                                iconRepo.refreshRetry(refreshWidget.getModpack());
-                                refreshWidget.updateFromPack(iconRepo.startImageJob(refreshWidget.getModpack()));
-
-                                EventQueue.invokeLater(new Runnable() {
-                                    @Override
-                                    public void run() {
-                                        revalidate();
-                                        repaint();
-                                    }
-                                });
+                            if (refreshWidget.getModpack().hasRecommendedUpdate()) {
+                                refreshWidget.setToolTipText(resources.getString("launcher.packselector.updatetip"));
+                            } else {
+                                refreshWidget.setToolTipText(null);
                             }
+
+                            iconRepo.refreshRetry(refreshWidget.getModpack());
+                            refreshWidget.updateFromPack(iconRepo.startImageJob(refreshWidget.getModpack()));
+
+                            EventQueue.invokeLater(() -> {
+                                revalidate();
+                                repaint();
+                            });
                         });
 
                     } catch (RestfulAPIException ex) {
@@ -377,12 +366,7 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
 
     protected void rebuildUI() {
         if (!EventQueue.isDispatchThread()) {
-            EventQueue.invokeLater(new Runnable() {
-                @Override
-                public void run() {
-                    rebuildUI();
-                }
-            });
+            EventQueue.invokeLater(() -> rebuildUI());
             return;
         }
 
@@ -390,21 +374,18 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
 
         java.util.List<ModpackWidget> sortedPacks = new LinkedList<ModpackWidget>();
         sortedPacks.addAll(allModpacks.values());
-        Collections.sort(sortedPacks, new Comparator<ModpackWidget>() {
-            @Override
-            public int compare(ModpackWidget o1, ModpackWidget o2) {
-                int priorityCompare = (new Integer(o2.getModpack().getPriority())).compareTo(new Integer(o1.getModpack().getPriority()));
-                if (priorityCompare != 0)
-                    return priorityCompare;
-                else if (o1.getModpack().getDisplayName() == null && o2.getModpack().getDisplayName() == null)
-                    return 0;
-                else if (o1.getModpack().getDisplayName() == null)
-                    return -1;
-                else if (o2.getModpack().getDisplayName() == null)
-                    return 1;
-                else
-                    return o1.getModpack().getDisplayName().compareToIgnoreCase(o2.getModpack().getDisplayName());
-            }
+        Collections.sort(sortedPacks, (o1, o2) -> {
+            int priorityCompare = (new Integer(o2.getModpack().getPriority())).compareTo(new Integer(o1.getModpack().getPriority()));
+            if (priorityCompare != 0)
+                return priorityCompare;
+            else if (o1.getModpack().getDisplayName() == null && o2.getModpack().getDisplayName() == null)
+                return 0;
+            else if (o1.getModpack().getDisplayName() == null)
+                return -1;
+            else if (o2.getModpack().getDisplayName() == null)
+                return 1;
+            else
+                return o1.getModpack().getDisplayName().compareToIgnoreCase(o2.getModpack().getDisplayName());
         });
 
         widgetList.removeAll();
@@ -424,12 +405,9 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
         constraints.weighty = 1.0;
         widgetList.add(Box.createGlue(), constraints);
 
-        EventQueue.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                revalidate();
-                repaint();
-            }
+        EventQueue.invokeLater(() -> {
+            revalidate();
+            repaint();
         });
     }
 
@@ -499,62 +477,78 @@ public class ModpackSelector extends TintablePanel implements IModpackContainer,
         setTintActive(true);
         defaultPacks.removePassthroughContainer(this);
 
-        currentSearchTimer = new Timer(500, new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String localSearchTag = searchText.trim();
+        currentSearchTimer = new Timer(500, e -> {
+            String localSearchTag = searchText.trim();
 
-                String localSearchUrl = searchText.trim();
-                if (!localSearchUrl.startsWith("http://") && !localSearchUrl.startsWith("https://"))
-                    localSearchUrl = "https://" + localSearchTag;
+            String localSearchUrl = searchText.trim();
+            if (!localSearchUrl.startsWith("http://") && !localSearchUrl.startsWith("https://"))
+                localSearchUrl = "https://" + localSearchTag;
 
-                try {
-                    URI uri = new URI(localSearchUrl);
-                    String host = uri.getHost();
-                    String scheme = uri.getScheme();
-                    if (host != null && scheme != null && (scheme.equals("http") || scheme.equals("https")) && (host.equals("www.technicpack.net") || host.equals("technicpack.net") || host.equals("api.technicpack.net"))) {
-                        String path = uri.getPath();
-                        if (path.startsWith("/"))
-                            path = path.substring(1);
-                        if (path.endsWith("/"))
-                            path = path.substring(0, path.length()-1);
-                        String[] fragments = path.split("/");
+            try {
+                URI uri = new URI(localSearchUrl);
+                String host = uri.getHost();
+                String scheme = uri.getScheme();
+                if (host != null && scheme != null && (scheme.equals("http") || scheme.equals("https"))) {
+                    String path = uri.getPath();
+                    if (path.startsWith("/"))
+                        path = path.substring(1);
+                    if (path.endsWith("/"))
+                        path = path.substring(0, path.length()-1);
+                    String[] fragments = path.split("/");
 
-                        if ((fragments.length == 2 && fragments[0].equals("modpack")) || (fragments.length == 3 && fragments[0].equals("api") && fragments[1].equals("modpack"))) {
-                            String slug = fragments[fragments.length-1];
+                    if ((fragments.length == 2 && fragments[0].equals("modpack")) || (fragments.length == 3 && fragments[1].equals("modpack"))) {
+                        String slug = fragments[fragments.length-1];
 
-                            Matcher siteMatcher = siteRegex.matcher(slug);
-                            if (siteMatcher.find()) {
-                                slug = siteMatcher.group(1);
-                            }
+                        Matcher siteMatcher = siteRegex.matcher(slug);
+                        if (siteMatcher.find()) {
+                            slug = siteMatcher.group(1);
+                        }
 
-                            Matcher slugMatcher = slugRegex.matcher(slug);
-                            if (slugMatcher.find()) {
-                                findMoreUrl = localSearchUrl;
-                                findMoreWidget.setWidgetData(resources.getString("launcher.packselector.api"));
-                                ArrayList<IPackSource> source = new ArrayList<IPackSource>(1);
-                                source.add(new SinglePlatformSource(platformApi, solderApi, slug));
-                                currentLoadJob = packLoader.createRepositoryLoadJob(ModpackSelector.this, source, null, false);
-                                return;
-                            }
+                        Matcher slugMatcher = slugRegex.matcher(slug);
+                        if (slugMatcher.find()) {
+                            Utils.getLogger().warning("Http Search Site: " + scheme + host + "/" + " Slug: " +slug + " Search text: "+searchText);
+                            IPlatformApi tempPlatform = new HttpPlatformApi(scheme + "://" + host + "/");
+                            findMoreUrl = localSearchUrl;
+                            findMoreWidget.setWidgetData(resources.getString("launcher.packselector.api"));
+                            ArrayList<IPackSource> source = new ArrayList<IPackSource>(1);
+                            source.add(new SinglePlatformSource(tempPlatform, solderApi, slug));
+                            currentLoadJob = packLoader.createRepositoryLoadJob(ModpackSelector.this, source, null, false);
+                            return;
                         }
                     }
-                } catch (URISyntaxException ex) {
-                    //It wasn't a valid URI which is actually fine.
                 }
-
-                String encodedSearch = filterContents.getText();
-                try {
-                    encodedSearch = URLEncoder.encode(encodedSearch, "UTF-8");
-                } catch (UnsupportedEncodingException ex) {}
-                findMoreUrl = "https://www.technicpack.net/modpacks?q="+encodedSearch;
-                findMoreWidget.setWidgetData(resources.getString("launcher.packselector.more"));
-
-                ArrayList<IPackSource> sources = new ArrayList<IPackSource>(2);
-                sources.add(new NameFilterPackSource(defaultPacks, localSearchTag));
-                sources.add(new SearchResultPackSource(platformSearchApi, localSearchTag));
-                currentLoadJob = packLoader.createRepositoryLoadJob(ModpackSelector.this, sources, null, false);
+            } catch (URISyntaxException ex) {
+                //It wasn't a valid URI which is actually fine.
             }
+
+            String encodedSearch;
+            ArrayList<IPackSource> sources = new ArrayList<>(2);
+            IPlatformSearchApi tempSearchapi = platformSearchApi;
+            PackLoader temploader = packLoader;
+            Utils.getLogger().warning(localSearchTag);
+            Utils.getLogger().log(Level.SEVERE, "Hell " + String.valueOf(localSearchTag.charAt(0)));
+            Utils.getLogger().config("hell2: " + (localSearchTag.charAt(0) == '`'));
+            if(localSearchTag.charAt(0) == '`'){
+                IPlatformApi tempPlatform = new HttpPlatformApi();
+                IAuthoritativePackSource tempauth = new PlatformPackInfoRepository(tempPlatform, solderApi);
+                tempSearchapi = new HttpPlatformSearchApi();
+
+                temploader = new PackLoader(temploader, tempauth);
+                localSearchTag = localSearchTag.substring(1);
+                Utils.getLogger().config("Hello3 " +localSearchTag);
+                encodedSearch  = filterContents.getText().substring(1);
+            } else {
+                sources.add(new NameFilterPackSource(defaultPacks, localSearchTag));
+                encodedSearch  = filterContents.getText();
+            }
+            try {
+                encodedSearch = URLEncoder.encode(encodedSearch, "UTF-8");
+            } catch (UnsupportedEncodingException ignored) {}
+            findMoreUrl = "https://www.technicpack.net/modpacks?q="+encodedSearch;
+            findMoreWidget.setWidgetData(resources.getString("launcher.packselector.more"));
+
+            sources.add(new SearchResultPackSource(tempSearchapi, localSearchTag));
+            currentLoadJob = temploader.createRepositoryLoadJob(ModpackSelector.this, sources, null, false);
         });
         currentSearchTimer.setRepeats(false);
         currentSearchTimer.start();
