@@ -105,12 +105,6 @@ public class ModpackModel {
         }
     }
 
-    public String getDiscordId() {
-        if (packInfo != null)
-            return packInfo.getDiscordId();
-        else
-            return null;
-    }
 
     public String getName() {
         if (packInfo != null) {
@@ -434,11 +428,15 @@ public class ModpackModel {
 
     public void save() {
         if (installedPack == null) {
-            installedPack = new InstalledPack(getName(), getBuild());
+            installedPack = new InstalledPack(getName(), getBuild(), getSource());
         }
 
         installedPackRepository.put(installedPack);
         installedPackRepository.save();
+    }
+
+    private String getSource() {
+    return null;
     }
 
     public boolean isSelected() {
